requirements.txt
Telethon
tgcrypto
Pyrogram==2.0.93
psutil
opencv-python-headless
python-decouple
requests
speedtest-cli
pymongo
yt-dlp
